# contracts/openapi/scilab.yaml
openapi: 3.1.0
info:
  title: SciLab REST API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
  /v1/me:
    get:
      operationId: getMe
      description: Verified bearer principal and its active Lab-scoped permissions.
      responses:
        '200':
          description: Current identity
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Me'}
  /v1/labs/{lab}/members:
    parameters:
      - $ref: '#/components/parameters/Lab'
    get:
      operationId: listLabMembers
      x-required-scopes: [lab:admin]
      responses:
        '200':
          description: Lab members
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MemberList'}
    post:
      operationId: addLabMember
      x-required-scopes: [lab:admin]
      description: Human Lab owner adds an existing OIDC subject; does not provision an account.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Member'}
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
  /v1/labs/{lab}/members/{subject}:
    parameters:
      - $ref: '#/components/parameters/Lab'
      - name: subject
        in: path
        required: true
        schema: {type: string, minLength: 1}
    patch:
      operationId: changeLabMemberRole
      x-required-scopes: [lab:admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MemberRole'}
      responses:
        '200':
          description: Role changed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
        '409': {description: Cannot demote the final owner}
    delete:
      operationId: removeLabMember
      x-required-scopes: [lab:admin]
      responses:
        '204': {description: Member removed}
        '409': {description: Cannot remove the final owner}
  /v1/labs/{lab}/budget:
    parameters:
      - $ref: '#/components/parameters/Lab'
    get:
      operationId: getLabBudget
      x-required-scopes: [lab:admin]
      responses:
        '200':
          description: Configured THB budget, or null when not configured
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LabBudget'}
    put:
      operationId: setLabBudget
      x-required-scopes: [lab:admin]
      description: Human Lab owner sets or clears the THB budget.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LabBudgetInput'}
      responses:
        '200':
          description: Budget updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LabBudget'}
  /v1/labs/{lab}/runs:
    parameters:
      - $ref: '#/components/parameters/Lab'
    post:
      operationId: createRun
      x-required-scopes: [runs:write]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RunRequest'}
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Object'}
    get:
      operationId: listRuns
      x-required-scopes: [runs:read]
      parameters:
        - {name: state, in: query, schema: {type: string}}
        - {name: actor, in: query, schema: {type: string}}
        - {name: since, in: query, schema: {type: string, format: date-time}}
        - {name: cursor, in: query, schema: {type: string}}
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getRun
      x-required-scopes: [runs:read]
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/runs/{id}/events:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: streamRunEvents
      x-required-scopes: [runs:read]
      parameters:
        - name: from_seq
          in: query
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: Resumable SSE stream with a heartbeat every 15 seconds
          content:
            text/event-stream:
              schema: {type: string}
  /v1/runs/{id}/stop:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: stopRun
      x-required-scopes: [runs:write]
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/runs/{id}/approvals/{approval_id}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - name: approval_id
        in: path
        required: true
        schema: {type: string}
    post:
      operationId: decideApproval
      x-required-scopes: [runs:approve]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalRequest'}
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/runs/{id}/artifacts:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: listRunArtifacts
      x-required-scopes: [artifacts:read]
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/artifacts/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getArtifact
      description: Returns a presigned URL valid for 15 minutes.
      x-required-scopes: [artifacts:read]
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/artifacts/{id}/content:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getArtifactContent
      x-required-scopes: [artifacts:read]
      description: Tenant-checked artifact bytes for report, manifest and image previews.
      responses:
        '200':
          description: Verified artifact bytes
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
  /v1/labs/{lab}/inputs:
    parameters:
      - $ref: '#/components/parameters/Lab'
    post:
      operationId: uploadInput
      x-required-scopes: [runs:write]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/labs/{lab}/ask:
    parameters:
      - $ref: '#/components/parameters/Lab'
    post:
      operationId: askLab
      description: Stateless Hermes chat; does not open a sandbox.
      x-required-scopes: [runs:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [question]
              properties:
                question: {type: string, minLength: 1}
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/labs/{lab}/skills:
    parameters:
      - $ref: '#/components/parameters/Lab'
    get:
      operationId: listSkills
      x-required-scopes: [runs:read]
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/labs/{lab}/api-keys:
    parameters:
      - $ref: '#/components/parameters/Lab'
    get:
      operationId: listApiKeys
      x-required-scopes: [lab:admin]
      responses:
        '200': {$ref: '#/components/responses/Json'}
    post:
      operationId: createApiKey
      x-required-scopes: [lab:admin]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses:
        '200': {$ref: '#/components/responses/Json'}
    delete:
      operationId: deleteApiKey
      x-required-scopes: [lab:admin]
      parameters:
        - name: key_id
          in: query
          required: true
          schema: {type: string, minLength: 1}
      responses:
        '204':
          description: Key deleted
  /v1/labs/{lab}/peers:
    parameters:
      - $ref: '#/components/parameters/Lab'
    get:
      operationId: listPeers
      x-required-scopes: [lab:admin]
      responses:
        '200': {$ref: '#/components/responses/Json'}
    post:
      operationId: createPeer
      x-required-scopes: [lab:admin]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses:
        '200': {$ref: '#/components/responses/Json'}
  /v1/labs/{lab}/usage:
    parameters:
      - $ref: '#/components/parameters/Lab'
    get:
      operationId: getUsage
      x-required-scopes: [runs:read]
      parameters:
        - name: period
          in: query
          required: true
          schema: {type: string, enum: [daily, monthly]}
      responses:
        '200': {$ref: '#/components/responses/Json'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Lab:
      name: lab
      in: path
      required: true
      schema: {type: string, minLength: 1}
    Id:
      name: id
      in: path
      required: true
      schema: {type: string, minLength: 1}
  schemas:
    Me:
      type: object
      additionalProperties: false
      required: [lab_id, principal, scopes]
      properties:
        lab_id: {type: string}
        principal: {type: string}
        scopes: {type: array, items: {type: string}}
    Member:
      type: object
      additionalProperties: false
      required: [subject, role]
      properties:
        subject: {type: string, minLength: 1}
        role: {type: string, enum: [owner, researcher, viewer]}
    MemberRole:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: {type: string, enum: [owner, researcher, viewer]}
    MemberList:
      type: object
      additionalProperties: false
      required: [members]
      properties:
        members: {type: array, items: {$ref: '#/components/schemas/Member'}}
    LabBudget:
      type: object
      additionalProperties: false
      required: [budget_thb]
      properties:
        budget_thb:
          description: Decimal THB encoded as a string to preserve precision; null means unconfigured.
          type: [string, 'null']
          pattern: '^[0-9]+(\\.[0-9]+)?$'
    LabBudgetInput:
      type: object
      additionalProperties: false
      required: [budget_thb]
      properties:
        budget_thb:
          oneOf:
            - {type: number, minimum: 0}
            - {type: string, pattern: '^[0-9]+(\\.[0-9]+)?$'}
            - {type: 'null'}
    Object:
      type: object
      additionalProperties: true
    RunRequest:
      type: object
      additionalProperties: false
      required: [goal, inputs, skill_packs, budget]
      properties:
        goal: {type: string, minLength: 1}
        inputs:
          type: array
          items: {type: string}
        skill_packs:
          type: array
          items: {type: string}
        budget:
          type: object
          additionalProperties: false
          required: [thb, max_minutes]
          properties:
            thb: {type: number, minimum: 0}
            max_minutes: {type: integer, minimum: 1}
        options:
          type: object
          additionalProperties: true
          default: {}
    ApprovalRequest:
      type: object
      additionalProperties: false
      required: [decision]
      properties:
        decision: {type: string, enum: [approve, reject]}
        note: {type: [string, 'null']}
  requestBodies:
    Object:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Object'}
  responses:
    Json:
      description: Successful response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Object'}